+++ testData/org/elixir_lang/beam/decompiler/Docs/Elixir.Kernel.ex
@@ -503,8 +503,8 @@ defmodule Kernel do
     )
   end
 
-  defmacro def(x0) do
-    super(x0, nil)
+  defmacro def(call) do
+    super(call, nil)
   end
 
   @doc ~S"""
@@ -774,8 +774,8 @@ defmodule Kernel do
     define_guard(:defmacrop, guard, __CALLER__)
   end
 
-  defmacro defimpl(x0, x1) do
-    super(x0, x1, [])
+  defmacro defimpl(name, opts) do
+    super(name, opts, [])
   end
 
   @doc ~S"""
@@ -792,8 +792,8 @@ defmodule Kernel do
     )
   end
 
-  defmacro defmacro(x0) do
-    super(x0, nil)
+  defmacro defmacro(call) do
+    super(call, nil)
   end
 
   @doc ~S"""
@@ -825,8 +825,8 @@ defmodule Kernel do
     define(:defmacro, call, expr, __CALLER__)
   end
 
-  defmacro defmacrop(x0) do
-    super(x0, nil)
+  defmacro defmacrop(call) do
+    super(call, nil)
   end
 
   @doc ~S"""
@@ -1005,8 +1005,8 @@ defmodule Kernel do
     {{:".", [], [{:__aliases__, [alias: false], [:"Module"]}, :make_overridable]}, [], [{:__MODULE__, [], Kernel}, keywords_or_behaviour]}
   end
 
-  defmacro defp(x0) do
-    super(x0, nil)
+  defmacro defp(call) do
+    super(call, nil)
   end
 
   @doc ~S"""
@@ -2280,8 +2280,8 @@ defmodule Kernel do
     end
   end
 
-  defmacro use(x0) do
-    super(x0, [])
+  defmacro use(module) do
+    super(module, [])
   end
 
   @doc ~S"""
@@ -3452,8 +3452,8 @@ defmodule Kernel do
     :erlang.hd(list)
   end
 
-  def inspect(x0) do
-    super(x0, [])
+  def inspect(term) do
+    super(term, [])
   end
 
   @doc ~S"""
@@ -4220,8 +4220,8 @@ defmodule Kernel do
     :erlang.spawn_monitor(module, fun, args)
   end
 
-  def struct(x0) do
-    super(x0, [])
+  def struct(struct) do
+    super(struct, [])
   end
 
   @doc ~S"""
@@ -4280,8 +4280,8 @@ defmodule Kernel do
     end)
   end
 
-  def struct!(x0) do
-    super(x0, [])
+  def struct!(struct) do
+    super(struct, [])
   end
 
   @doc ~S"""